Extract admin gate check for isolated testing
